# docs/openapi.generated.yaml
openapi: 3.1.0
info:
  title: Racing Point racecontrol API
  description: Generated from utoipa annotations. Admin-surface subset (Phase 445 D-06).
  contact:
    name: RacingPoint
    email: bono@racingpoint.in
  license:
    name: MIT
    identifier: MIT
  version: 1.0.0
paths:
...
  /api/v1/billing/rates:
    get:
      tags:
      - billing
      operationId: list_billing_rates
      responses:
        '200':
          description: List of per-minute billing rate tiers
          content:
            application/json:
              schema: {}
        '401':
          description: Staff JWT required
      security:
      - staffJWT: []
...
  /api/v1/pricing:
    get:
      tags:
      - pricing
      operationId: list_pricing_tiers
      responses:
        '200':
          description: List of pricing tiers
          content:
            application/json:
              schema: {}
        '401':
          description: Staff JWT required
      security:
      - staffJWT: []
...